# Conflicts:
#	order/listener/src/main/kotlin/com/rarible/protocol/order/listener/configuration/OrderListenerConfiguration.kt

+++ order/listener/src/test/kotlin/com/rarible/protocol/order/listener/job/ReservoirOrdersReconciliationWorkerTest.kt
@@ -142,7 +142,7 @@ internal class ReservoirOrdersReconciliationWorkerTest {
                 ),
                 amount = Amount(
                     raw = randomBigInt(),
-                    decimal = 10,
+                    decimal = 10.toBigDecimal(),
                     usd = randomBigDecimal(),
                     native = randomBigDecimal(),
                 ),
